software/autonomy_ws/src/vlm_driver/requirements.txt
transformers>=4.45
accelerate
huggingface_hub
torch
pillow
pyyaml
qwen-vl-utils[decord]
hf_transfer